+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@ This is the offical release of vpinfe v1.1.62
 
 ### What's New
 - VPX config editor filter by non-defaults
+- VPX config editor add backup/restore
 
 ### Fixes
 None